requirements.txt
auth0-python
cryptography
fastapi
pydantic
python-dotenv
PyJWT
pymongo
secure
uvicorn